+++ .github/workflows/run.yml
@@ -133,4 +133,7 @@ jobs:
           pip install -r ./requirements.txt
       - name: Run
         run: |-
-          ./bin/benchpark unit-test
+          for file in lib/benchpark/test/*.py; do
+            echo "Running tests in $file"
+            ./bin/benchpark unit-test "$file"
+          done
